feat(daterangepicker component): add "This Week" and "Next Week" buttons
adds two new buttons, "This Week" and "Next Week", to the DateRangePicker component. The "This Week"
button sets the date range to the current week, starting from Monday and ending on Sunday. The "Next
Week" button sets the date range to the following week, also starting from Monday and ending on
Sunday.