refactor(events page): refactor location filtering to use new data source
refactors the location filtering in the Events component to use a new data source for the list of
locations. Previously, the list of locations was hard-coded in an array, but it has now been moved
to a separate file. The YemenCities array is now imported and used instead of the locations array